Conversion from sequential to modular programming structure and some modifications
I converted the sequential programming structure into modular programming, I added function for the operations performed by the manager so that the code looks easier to understand.
I used the pass by reference in the functions which I added.
There were also some faults in the previous code such as it couldn't display the message of invalid registration number in the chunk of code for printing the student file.